Alerts

+++ components/FavouritesComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
@@ -34,7 +34,25 @@ class Favourites extends Component {
                 {
                     text: 'Delete', 
                     type: 'delete',
-                    onPress: () => this.props.deleteFavourite(item.id)
+                    onPress: () => {
+                        Alert.alert (
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favourite dish ' + item.name + '?',
+                            [
+                                { 
+                                    text: 'Cancel', 
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavourite(item.id)
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                    }
+                    
                 }
             ];
